pyproject.toml
[tool.poetry]
name = "oz-django-study"
version = "0.1.0"
description = "python django project"
authors = ["SOLBI <thfqlsla35@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.4"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

